backend/requirements.txt
flask
flask-cors
supabase
python-dotenv
pandas
numpy
